# docs/openapi.yaml
openapi: 3.0.3
info:
  title: FITS Backend API
  version: 1.0.0
  description: |
    # FITS (Flexible IT Training System) Backend API

    Production-ready REST API for managing students, teachers, authentication, and signing workflows.

    ## Features
    - JWT-based authentication with refresh tokens
    - Role-based access control (Admin, Teacher, Student)
    - Invitation system for user onboarding
    - Student and teacher management with pagination
    - Digital signing workflow for reports
    - Rate limiting and security headers

    ## Authentication Flow
    1. **Bootstrap Admin**: `POST /api/v1/bootstrap/init` (one-time setup)
    2. **Login**: `POST /api/v1/auth/login` (get access + refresh tokens)
    3. **Use Access Token**: Include `Authorization: Bearer {access_token}` header
    4. **Refresh Token**: `POST /api/v1/auth/refresh` when access token expires
    5. **Logout**: `POST /api/v1/auth/logout` to invalidate tokens

    ## Rate Limiting
    - Global: 100 requests/minute per IP
    - Per-user: Varies by role (Admin: 200/min, Teacher: 100/min, Student: 50/min)

    ## Pagination
    All list endpoints support:
    - `page` (default: 1, min: 1)
    - `limit` (default: 20, max: 100)
  termsOfService: https://fits.example.com/terms
  contact:
    name: API Support Team
    email: support@fits.example.com
    url: https://fits.example.com/support
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/auth/logout:
    post:
      summary: User logout
      description: |
        Invalidate all refresh tokens for the authenticated user.
        Access tokens remain valid until expiry (cannot be revoked).
      operationId: authLogout
      tags: [Authentication]
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
              examples:
                success:
                  summary: Successfully logged out
                  value:
                    success: true
                    message: Logged out successfully
                    data: null
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== INVITATIONS ====================
  /api/v1/admin/invite:
    post:
      summary: Create invitation
      description: |
        Create invitation link for new student or teacher (Admin only).

        **Required fields by role:**
        - **Student**: email, first_name, last_name, role=student, teacher_uuid
        - **Teacher**: email, first_name, last_name, role=teacher, department
      operationId: createInvitation
      tags: [Invitations]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
            examples:
              studentInvite:
                summary: Invite a student
                value:
                  email: max.mustermann@example.com
                  first_name: Max
                  last_name: Mustermann
                  role: student
                  teacher_uuid: 550e8400-e29b-41d4-a716-446655440010
              teacherInvite:
                summary: Invite a teacher
                value:
                  email: anna.schmidt@example.com
                  first_name: Anna
                  last_name: Schmidt
                  role: teacher
                  department: Computer Science
      responses:
        '201':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CreateInvitationResponse'
              examples:
                success:
                  summary: Invitation created
                  value:
                    success: true
                    message: Invitation created successfully
                    data:
                      invitation_token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                      invitation_link: https://fits.example.com/invite/eyJhbGc...
                      expires_at: "2025-10-25T12:00:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: Validation error - missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                missingTeacherUuid:
                  summary: Student missing teacher_uuid
                  value:
                    success: false
                    code: 422
                    error: Validation Error
                    details: "teacher_uuid is required for student invitations"
                missingDepartment:
                  summary: Teacher missing department
                  value:
                    success: false
                    code: 422
                    error: Validation Error
                    details: "department is required for teacher invitations"
...
  /api/v1/student:
    get:
      summary: List students with pagination
      description: |
        Retrieve a paginated list of students.
        Supports filtering via query parameters.
      operationId: listStudents
      tags: [Students]
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of students
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedStudentsResponse'
              examples:
                success:
                  summary: First page with 2 students
                  value:
                    success: true
                    page: 1
                    limit: 20
                    total_count: 150
                    total_pages: 8
                    data:
                      - uuid: 550e8400-e29b-41d4-a716-446655440000
                        email: max.mustermann@example.com
                        first_name: Max
                        last_name: Mustermann
                        teacher_id: 550e8400-e29b-41d4-a716-446655440010
                        created_at: "2025-09-30T12:00:00Z"
                        updated_at: "2025-09-30T12:00:00Z"
                      - uuid: 550e8400-e29b-41d4-a716-446655440001
                        email: lisa.mueller@example.com
                        first_name: Lisa
                        last_name: Mueller
                        teacher_id: 550e8400-e29b-41d4-a716-446655440010
                        created_at: "2025-09-29T10:30:00Z"
                        updated_at: "2025-09-29T10:30:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new student
      description: |
        Create a new student record. Requires admin role.
        Email must be unique.
      operationId: createStudent
      tags: [Students]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStudentRequest'
            examples:
              withTeacher:
                summary: Student with assigned teacher
                value:
                  email: max.mustermann@example.com
                  first_name: Max
                  last_name: Mustermann
                  teacher_id: 550e8400-e29b-41d4-a716-446655440010
              withoutTeacher:
                summary: Student without teacher (assign later)
                value:
                  email: lisa.mueller@example.com
                  first_name: Lisa
                  last_name: Mueller
      responses:
        '201':
          description: Student created successfully
          headers:
            Location:
              schema:
                type: string
              description: URI of created student
              example: /api/v1/student/550e8400-e29b-41d4-a716-446655440000
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Student'
              examples:
                success:
                  value:
                    success: true
                    message: Student created successfully
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440000
                      email: max.mustermann@example.com
                      first_name: Max
                      last_name: Mustermann
                      teacher_id: 550e8400-e29b-41d4-a716-446655440010
                      created_at: "2025-10-22T14:30:00Z"
                      updated_at: "2025-10-22T14:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailConflict:
                  value:
                    success: false
                    code: 409
                    error: Conflict
                    details: A student with this email already exists
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/teacher:
    get:
      summary: List teachers with pagination
      description: Retrieve a paginated list of teachers
      operationId: listTeachers
      tags: [Teachers]
      security: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: Paginated list of teachers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTeachersResponse'
              examples:
                success:
                  value:
                    success: true
                    page: 1
                    limit: 20
                    total_count: 45
                    total_pages: 3
                    data:
                      - uuid: 550e8400-e29b-41d4-a716-446655440010
                        email: anna.schmidt@example.com
                        first_name: Anna
                        last_name: Schmidt
                        department: Computer Science
                        created_at: "2025-09-28T10:00:00Z"
                        updated_at: "2025-09-28T10:00:00Z"
        '500':
          $ref: '#/components/responses/InternalServerError'

    post:
      summary: Create a new teacher
      description: |
        Create a new teacher record. Requires admin role.
        Email must be unique. Department is required.
      operationId: createTeacher
      tags: [Teachers]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTeacherRequest'
            examples:
              computerScience:
                summary: Computer Science teacher
                value:
                  email: anna.schmidt@example.com
                  first_name: Anna
                  last_name: Schmidt
                  department: Computer Science
              mathematics:
                summary: Mathematics teacher
                value:
                  email: peter.mueller@example.com
                  first_name: Peter
                  last_name: Mueller
                  department: Mathematics
      responses:
        '201':
          description: Teacher created successfully
          headers:
            Location:
              schema:
                type: string
              description: URI of created teacher
              example: /api/v1/teacher/550e8400-e29b-41d4-a716-446655440010
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Teacher'
              examples:
                success:
                  value:
                    success: true
                    message: Teacher created successfully
                    data:
                      uuid: 550e8400-e29b-41d4-a716-446655440010
                      email: anna.schmidt@example.com
                      first_name: Anna
                      last_name: Schmidt
                      department: Computer Science
                      created_at: "2025-10-22T14:30:00Z"
                      updated_at: "2025-10-22T14:30:00Z"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                emailConflict:
                  value:
                    success: false
                    code: 409
                    error: Conflict
                    details: A teacher with this email already exists
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Bearer token authentication.

        **Format**: `Authorization: Bearer {access_token}`

        **How to get tokens:**
        1. Login: `POST /api/v1/auth/login`
        2. Use returned `access_token` in Authorization header
        3. Refresh when expired: `POST /api/v1/auth/refresh`
  parameters:
    PageParam:
      name: page
      in: query
      required: false
      description: Page number (1-indexed)
      schema:
        type: integer
        minimum: 1
        default: 1
      example: 1
    LimitParam:
      name: limit
      in: query
      required: false
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      example: 20
...
  responses:
    BadRequest:
      description: Bad request - invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            invalidJson:
              summary: Invalid JSON
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid JSON format
            invalidUuid:
              summary: Invalid UUID format
              value:
                success: false
                code: 400
                error: Bad Request
                details: Invalid UUID format
    Unauthorized:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            missingToken:
              summary: Missing authorization header
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Missing authorization token
            invalidToken:
              summary: Invalid or expired token
              value:
                success: false
                code: 401
                error: Unauthorized
                details: Invalid or expired token
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            insufficientPermissions:
              summary: Admin role required
              value:
                success: false
                code: 403
                error: Forbidden
                details: Admin role required
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          examples:
            serverError:
              summary: Generic server error
              value:
                success: false
                code: 500
                error: Internal Server Error
                details: An unexpected error occurred. Please try again later.
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operation successful
        data:
          type: object
          nullable: true
          description: Response payload (varies by endpoint)
    ErrorResponse:
      type: object
      required: [success, code, error]
      properties:
        success:
          type: boolean
          example: false
        code:
          type: integer
          description: HTTP status code
          example: 400
        error:
          type: string
          description: Error type/category
          example: Bad Request
        details:
          type: string
          description: Human-readable error details
          example: Invalid email format
...
    CreateInvitationRequest:
      type: object
      required: [email, first_name, last_name, role]
      properties:
        email:
          type: string
          format: email
          description: User's email address
          example: max.mustermann@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: User's first name
          example: Max
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: User's last name
          example: Mustermann
        role:
          type: string
          enum: [student, teacher]
          description: User's role
          example: student
        teacher_uuid:
          type: string
          format: uuid
          description: Required for student invitations - UUID of assigned teacher
          example: 550e8400-e29b-41d4-a716-446655440010
        department:
          type: string
          minLength: 1
          maxLength: 100
          description: Required for teacher invitations - teacher's department
          example: Computer Science
    CreateInvitationResponse:
      type: object
      required: [invitation_token, invitation_link, expires_at]
      properties:
        invitation_token:
          type: string
          description: JWT invitation token
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        invitation_link:
          type: string
          format: uri
          description: Complete invitation URL to send to user
          example: https://fits.example.com/invite/eyJhbGc...
        expires_at:
          type: string
          format: date-time
          description: Token expiration timestamp
          example: "2025-10-25T12:00:00Z"
...
    CreateStudentRequest:
      type: object
      required: [email, first_name, last_name]
      properties:
        email:
          type: string
          format: email
          description: Student's email address (must be unique)
          example: max.mustermann@example.com
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Student's first name
          example: Max
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          description: Student's last name
          example: Mustermann
        teacher_id:
          type: string
          format: uuid
          nullable: true
          description: UUID of assigned teacher (optional - can be assigned later)
          example: 550e8400-e29b-41d4-a716-446655440010
...
    PaginatedStudentsResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Student'

    # ==================== TEACHERS ====================
...
    PaginatedTeachersResponse:
      allOf:
        - $ref: '#/components/schemas/PaginationMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items:
                $ref: '#/components/schemas/Teacher'

    # ==================== SIGNING ====================
...